em_research_chat/pyproject.toml
[tool.poetry]
name = "em_research_chat"
version = "0.1.0"
description = "Emerging Markets News Agentic Research"
authors = [
    "RRaffay"
]
packages = [
    { include = "em_research_chat" },
]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.13"
langgraph = "^0.2.0"
langchain_anthropic = "^0.2.0"
langchain_core = "^0.3.0"
langchain_openai = "^0.2.0"
tavily-python = "^0.4.0"
langchain_community = "^0.3.0"
beautifulsoup4 = "^4.12.3"
fastapi = {extras = ["standard"], version = "^0.114.2"}
langchain-experimental = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"